docs: add link to ixy.js thesis
Adds the link to the finished thesis of the ixy.js implementation. It has been finished for quite some time, but the link was never added in the past because the paper wasnt publically available anywhere. Now that it's part of the ixy.js repo, we can link to it.